+++ config/config.exs
@@ -1,0 +1,6 @@
+use Mix.Config
+
+config :sentry,
+  included_environments: [:prod]
+
+import_config "#{Mix.env}.exs"
