+++ config/test.exs
@@ -1,0 +1,4 @@
+use Mix.Config
+
+config :sentry,
+  environment_name: :test
